# swagger.yml
openapi: 3.0.1
info:
  title: Phrase Recorder API
  description: This is the API for the Phrase Recorder App.
  termsOfService: http://swagger.io/terms/
  contact:
    email: gdavidson1107@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://indigenous.io/v1
...
paths:
...
  /userPhrase:
    put:
      tags:
        - userPhrase
      summary: Edits a userPhrase in the db
      operationId: updateUserPhrase
      requestBody:
        description: UserPhrase object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPhrase'
        required: true
      responses:
        "200":
          description: Successful operation
          content: {}
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: UserPhrase not found
          content: {}
        "405":
          description: Validation exception
          content: {}

      x-codegen-request-body-name: userPhrase
      x-openapi-router-controller: swagger_server.controllers.UserPhraseController
    post:
      tags:
        - userPhrase
      summary: Saves a userPhrase to db
      operationId: addUserPhrase
      requestBody:
        description: UserPhrase object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPhrase'
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPhrase'
        "405":
          description: Validation exception
          content: {}
      x-codegen-request-body-name: userPhrase
      x-openapi-router-controller: swagger_server.controllers.UserPhraseController
  /file:
    post:
      summary: Saves a voice recording to bucket for a particular user
      operationId: saveAudioFileToBucket
      requestBody:
        description: UserPhraseWithAudioFile object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPhraseWithAudioFile'
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPhrase'
        "405":
          description: Validation exception
          content: { }
      x-codegen-request-body-name: userPhrase
      x-openapi-router-controller: swagger_server.controllers.BucketStorageController
  /file/{keypath}:
    post:
      summary: Gets file from bucket with given keypath
      operationId: getFileFromBucket
      parameters:
        - name: keypath
          in: path
          description: The path of the file you want to fetch from the bucket.
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-openapi-router-controller: swagger_server.controllers.BucketStorageController
...
  /phraseIds:
    post:
      tags:
        - phraseId
      summary: get All Remaining PhraseIds For A User And PhraseSet
      operationId: getAllRemainingPhraseIdsForAUserAndPhraseSet
      requestBody:
        description: PhraseRequest object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhraseRequest'
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  format: int64


        "400":
          description: Invalid ID supplied
          content: { }
        "404":
          description: Not found
          content: { }
        "405":
          description: Validation exception
          content: { }
      x-codegen-request-body-name: phraseRequest
      x-openapi-router-controller: swagger_server.controllers.UserPhraseCountController



  ## USER PHRASE COMMENT
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        project:
          type: string
      xml:
        name: User
    UserPhrase:
      type: object
      properties:
        id:
          type: integer
          format: int64
        phraseId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        dateTime:
          type: string
          format: date-time
        filepath:
          type: string
          format:  binary
        ## mp4, mp3, ogg vorbis, wav
        ## hex encoded binary string (b64)
        ## add param to tell what kind of file it is, so like (binaryString, wav)
      xml:
        name: UserPhrase
...
    UserPhraseWithAudioFile:
      type: object
      properties:
        userPhrase:
          $ref: '#/components/schemas/UserPhrase'
        file:
          type: string
          format: byte
      xml:
        name: UserPhraseWithAudioFile
    PhraseRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        phraseSetId:
          type: integer
          format: int64
      xml:
        name: PhraseRequest